Move easyadmin_dump to collector instead of a Twig Extension
